+++ database/post_user_play_info.go
@@ -2,7 +2,7 @@ package database
 
 import "log"
 
-func PostUserPlayInfo(dataID uint64, ownerID uint32, slot uint16) {
+func PostUserPlayInfo(dataID uint32, ownerID uint32, slot uint16) {
 	var err error
 	_, err = postgres.Exec(`INSERT INTO pretendo_badge_arcade.user_play_info(
 		data_id,
