+++ utility/datastore_post_param_to_free_play_data.go
@@ -11,5 +11,5 @@ func DataStorePostParamToFreePlayData(ownerID uint32, metaInfo *datastore.DataSt
 	createdTime := dateTime.Now()
 
 	// We are setting the PID as the Data ID, as it is easier to handle
-	database.PostFreePlayDataMetaInfo(uint64(ownerID), ownerID, metaInfo.MetaBinary, createdTime, metaInfo.Period, metaInfo.Flag)
+	database.PostFreePlayDataMetaInfo(ownerID, ownerID, metaInfo.MetaBinary, createdTime, metaInfo.Period, metaInfo.Flag)
 }
